replace physical hands anchors with modified anchors which can be used for interaction engine or physical hands
