+++ Packages/Tracking/Interaction Engine/Runtime/Scripts/UI/Anchors.meta
@@ -1,9 +1,8 @@
 fileFormatVersion: 2
-guid: 96a11d203300103469d5a455c242cd66
+guid: 633fa65b11b690547ad8d425f4e2e117
 folderAsset: yes
-timeCreated: 1493327449
-licenseType: Free
 DefaultImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
