+++ Packages/Tracking/Interaction Engine/Runtime/Ultraleap.Tracking.InteractionEngine.asmdef
@@ -1,8 +1,10 @@
 {
     "name": "Ultraleap.Tracking.InteractionEngine",
+    "rootNamespace": "",
     "references": [
         "GUID:2ee0542102ac1084e8d9a910885c1e81",
-        "GUID:cec9848704c49724192ff47f9b3e9156"
+        "GUID:cec9848704c49724192ff47f9b3e9156",
+        "GUID:5d2bbb4763a30e54ebe3c25d112247b1"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
