--- Packages/Tracking/Physical Hands/Runtime/Scripts/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 18e39389a982aab47aca4667c3b4bd79
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
