--- Packages/Tracking/Physical Hands/Runtime/Scripts/UI/PhysicalHandsAnchor.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 39acd8caefb2273428888e3b1917cd9a
-timeCreated: 1492722870
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
