--- Packages/Tracking/Physical Hands/Runtime/Scripts/UI/PhysicalHandsAnchorGroup.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 51a01d08e8536324aab227ef0736dad8
-timeCreated: 1492722870
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
