--- Packages/Tracking/Physical Hands/Runtime/Scripts/UI/PhysicalHandsAnchorableBehaviour.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: fea4c084a50d93f40a5316aada996264
-timeCreated: 1492722870
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
